IoT/requirements.txt
arandr==0.1.10
astroid==2.5.1
asttokens==2.0.4
automationhat==0.2.0
beautifulsoup4==4.9.3
bidict==0.22.0
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
distro==1.5.0
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.1.2
fourletterphat==0.1.0
gevent==21.12.0
gpiozero==1.6.2
greenlet==1.1.2
html5lib==1.1
HX711 @ git+https://github.com/bytedisciple/HX711.git@5e99ae2194fe334beb6bf24873482ed32a05b6fe#subdirectory=HX711_Python3
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
lazy-object-proxy==0.0.0
logilab-common==1.8.1
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
mfrc522==0.0.7
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mypy==0.812
mypy-extensions==0.4.3
netifaces==0.10.6
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
psutil==5.8.0
pycairo==1.16.2
pycups==2.0.1
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
pyOpenSSL==20.0.1
pyserial==3.5b0
pysmbc==1.0.23
python-apt==2.2.1
python-dotenv==0.20.0
python-engineio==3.14.2
python-engineio-client==1.1
python-socketio==4.6.0
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
responses==0.12.1
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.3.1
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
socketIO-client==0.7.2
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
thonny==3.3.14
toml==0.10.1
touchphat==0.0.1
twython==3.8.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
unicornhathd==0.0.4
urllib3==1.26.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
